# app/Http/Requests/StoreFolderRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;
use OpenApi\Attributes as OA;

#[OA\Schema(
    properties: [
        'folder_name' => new OA\Property(property: 'folder_name', description: 'Название дочерней папки', type: 'string', example: 'Documents'),
    ]
)]
class StoreFolderRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\Rule|array|string>
     */
    public function rules(): array
    {
        return [
            'folder_name' => ['required', 'string'],
        ];
    }
}